requirements.txt
beautifulsoup4==4.3.2
pyenchant==1.6.6
requests==2.5.1
wsgiref==0.1.2
